referrer-policy/generic/inheritance/iframe-inheritance-document-write.html

<!doctype html>
<title>Referrer Policy: iframes with document.write()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/make-html-script.js"></script>
<meta name="referrer" content="origin">
<div id="log"></div>
<script>
  let reportedReferrer = () => {
    return new Promise(resolve => {
      window.addEventListener("message", msg => resolve(msg.data.referrer));
    });
  };

  const iframe = document.createElement("iframe");
  promise_test(async t => {
    let referrer_1 = reportedReferrer();
    iframe.srcdoc = `<head><meta name="referrer" content="unsafe-url"></head>`
      + createScriptString(get_host_info().REMOTE_ORIGIN, location.origin + "/custom");
    document.body.appendChild(iframe);
    assert_equals(await referrer_1, self.origin + "/custom", "1");

    let referrer_2 = reportedReferrer();
    iframe.contentDocument.open();
    iframe.contentDocument.write(createScriptString(get_host_info().REMOTE_ORIGIN, location.origin + "/custom"));
    iframe.contentDocument.close();
    assert_equals(await referrer_2, self.origin + "/custom", "2");
  }, "document.open should not change the referrer policy of the opened document.");
</script>
